Decouple ProviderManager from Services
With RemoteProvidersStrategy.
