+++ PassepartoutLibrary/Sources/PassepartoutProviders/Strategies/RemoteProvidersStrategy.swift
@@ -1,0 +1,51 @@
+//
+//  RemoteProvidersStrategy.swift
+//  Passepartout
+//
+//  Created by Davide De Rosa on 5/24/23.
+//  Copyright (c) 2023 Davide De Rosa. All rights reserved.
+//
+//  https://github.com/passepartoutvpn
+//
+//  This file is part of Passepartout.
+//
+//  Passepartout is free software: you can redistribute it and/or modify
+//  it under the terms of the GNU General Public License as published by
+//  the Free Software Foundation, either version 3 of the License, or
+//  (at your option) any later version.
+//
+//  Passepartout is distributed in the hope that it will be useful,
+//  but WITHOUT ANY WARRANTY; without even the implied warranty of
+//  MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
+//  GNU General Public License for more details.
+//
+//  You should have received a copy of the GNU General Public License
+//  along with Passepartout.  If not, see <http://www.gnu.org/licenses/>.
+//
+
+import Combine
+import Foundation
+import PassepartoutCore
+
+public enum RemoteProvidersPriority {
+    case bundle
+
+    case remote
+
+    case remoteThenBundle
+}
+
+public protocol RemoteProvidersStrategy {
+    func saveIndex(
+        priority: RemoteProvidersPriority,
+        onFetch: @escaping () -> Void
+    ) -> AnyPublisher<Void, Error>
+
+    func saveProvider(
+        withName providerName: ProviderName,
+        vpnProtocol: VPNProtocolType,
+        lastUpdate: Date?,
+        priority: RemoteProvidersPriority,
+        onFetch: @escaping () -> Void
+    ) -> AnyPublisher<Void, Error>
+}
